Fix sync issue: save unhandled leaves and reapply after syncState.
This can occur when the socket subscibes to presence updates, and
leave events (that occur after the socket process has fetched the state)
are pushed to the client before the socket process can push the state to
the client. It's more likely to happen if the server process has to fetch
user data for the full presence state.